file estructure refactored
