+++ src/application/product/product.module.ts
@@ -1,8 +1,7 @@
 import { Module, forwardRef } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { TransactionModule } from './transaction.module';
-import { Product } from 'src/domain/entities/product.entity';
-import { productService } from 'src/application/services/product.service';
+import { Product } from 'src/domain/product/product.entity';
+import { productService } from 'src/domain/product/product.service';
 
 @Module({
   imports: [
