+++ src/domain/customer/customer.service.ts
@@ -1,14 +1,14 @@
 import { Injectable, Inject, forwardRef } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { Customer } from 'src/domain/customer/customer.entity';
 import { Repository } from 'typeorm';
-import { TransactionService } from './Transaction.service';
-import { TransactionDetail } from 'src/domain/entities/transaction-detail.entity';
+import { TransactionService } from '../transaction/Transaction.service';
 
 @Injectable()
-export class TransactionDetailService {
+export class CustomerService {
   constructor(
-    @InjectRepository(TransactionDetail)
-    private readonly customerRepository: Repository<TransactionDetail>,
+    @InjectRepository(Customer)
+    private readonly customerRepository: Repository<Customer>,
     @Inject(forwardRef(() => TransactionService)) 
     private readonly transactionService: TransactionService,
   ) {}
