+++ src/domain/payment/payment.entity.ts
@@ -7,7 +7,7 @@ import {
   UpdateDateColumn,
   JoinColumn,
 } from 'typeorm';
-import { Transaction } from './transaction.entity';
+import { Transaction } from '../transaction/transaction.entity';
 
 @Entity()
 export class Payment {
