+++ src/domain/payment/payment.service.ts
@@ -1,8 +1,8 @@
 import { Injectable, Inject, forwardRef } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { TransactionService } from './Transaction.service';
-import { Payment } from 'src/domain/entities/payment.entity';
+import { TransactionService } from '../transaction/Transaction.service';
+import { Payment } from 'src/domain/payment/payment.entity';
 
 @Injectable()
 export class paymentService {
